+++ search-repositories-github/src/components/Dashboard/index.jsx
@@ -6,7 +6,8 @@ import { DashboardContainer, MainContent } from './styles'
 /**
  * Componente Dashboard
  * contêm os outros componentes que aparecem em todas as páginas
- * @param children componente filho a ser renderizado dentro do dashboard
+ * @param {*} children componente filho a ser renderizado dentro do dashboard
+ * @returns Dashboard React Component
  */
 export function Dashboard({ children }) {
   return (
