+++ search-repositories-github/src/components/Header/index.jsx
@@ -8,6 +8,7 @@ import { useGithubData } from '../../hooks/DataContext'
  * Componente Header
  * contêm o input para a pesquisa dos repositórios. A comunicação com a api
  * para buscar o repositório é realizada pelo hook useGithubData com o valor do input.
+ * @returns Header React Component
  */
 export function Header() {
   const [ searchText, setSearchText ] = useState('')
